# Optimize SAX2Parser#get_namespace